C43944: Extra vertical line
Hello, @mlottner,
Localization team has reported source content issue that causes issue for localization. 
Description of the source issue: There is an extra vertical bar (5 instead of 4), which breaks formatting on LOC site and the content is missing on ENU site.
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.Many thanks in advance.