C39922: Missing linebreak at line 62
Hi  @arob98 , @GitHubber17 , 
This changes comes from this PR https://github.com/MicrosoftDocs/ATADocs-pr/pull/865 and this was accepted. 
Regarding the creation of some PRs in the Live Branch, we were having some issues with the links we received to the files to modify but we are aware of this now.

Mauricio.