C39924: Inconsistent format
Hello, @rkarlin,
Localization team has reported source content issue that causes issue for localization. 
Description:
There are new problems present in the source file. It is difficult to find fix in the source file, because I am not able to recognize how the content between lines 43 - 74 should actually look like and it seems to be really inconsistent. In my opinion it should be completely rewritten. 
For example:
Line 59 has the same bullet points as lines 60 and 61, while the line 59 (based on the content) should be superior compared to lines 60/61.
another example of inconsistency:
There are two spaces between numbering and strings in lines 43 and 48, while there is only one space in lines 71/72/74. 

Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.