Update general-setup.md
changed:
Under **Auto sign out**, specify the amount of time a session can remain inactive before the session is automatically signed out.

to:
Under **Auto sign out**, specify the amount of time a session can remain inactive before the session is automatically signed out. Please note that Defender for Cloud Apps also uses Azure Active Directory to determine the user's directory level inactivity timeout setting. If a user is configured in Azure Active Directory to never sign out when inactive, the same setting will apply in Defender for Cloud Apps as well.

based on the same language as in the note in  https://docs.microsoft.com/en-us/defender-cloud-apps/manage-admins#office-365-and-azure-ad-roles-with-access-to-defender-for-cloud-apps

-------
cc: @dcurwin