Update tvm-weaknesses.md
Modified content on the Breach icon, which originally incorrectly stated that it would be enabled when the vulnerability is detected on an endpoint. It did not mention that it also has to be associated with an alert/a device had to have been targeted with an exploit related to the vulnerability.