Update android-intune.md
Customer requested a note to be added to this document to reflect the note in Intune document here: https://learn.microsoft.com/en-us/mem/intune/apps/app-configuration-policies-use-android#use-the-configuration-designer

Value type: The data type of the configuration value. For string value types, you can optionally choose a variable or certificate profile as the value type. Note that once the policy is created, these value types will show as string.