Update defender-endpoint-trial-user-guide.md
I'm part of the MDE Care Engineering Team, The reason for adding the note is that we've been receiving numerous ICM's requests for offboarding scripts after customer licenses expire. Previously, we had a pipeline in place to handle this on an exceptional basis, but it has since been decommissioned, and we no longer have that capability. By including this note, we aim to proactively inform customers and make the information publicly available, encouraging them to offboard devices before their trial period ends.