Add note that the procedure requires tenant Entra ID P2
Clarify that this procedure requires Microsoft Entra ID P2. 

That's important to MSSPs, if they want to onboard a customer using this, their customer has to pay for ID P2 according to the licensing sheet, it does not say that the MSSP's licensing can be reused. That is an expensive proposition and one most customers won't go for.