Update recommended-settings-for-eop-and-office365.md
Reading through this, setting a notification address for Default Outbound Spam policy recommends not configured because of Alert policy. However, as of 1/13/25, this is enabled on our tenant, and SecureScore is asking for it to be configured. Not sure how to link to a specific SS entry, but the title is "Ensure Exchange Online Spam Policies are set to notify administrators"

Implementation tasks, copied: 

Check Send a copy of outbound messages that exceed these limits to these users and groups then enter the desired email addresses.
Check Notify these users and groups if a sender is blocked due to sending outbound spam then enter the desired email addresses.